neuroglancer.annotatons.precomputed: Split the implementation across multiple files for better organization
