+++ neuclease/misc/neuroglancer/annotations/precomputed/precomputed.py
@@ -6,8 +6,6 @@ from typing import Literal
 
 import pandas as pd
 import numpy as np
-from tqdm import tqdm
-import tensorstore as ts
 
 try:
     from neuroglancer.coordinate_space import CoordinateSpace
@@ -21,7 +19,12 @@ except ImportError:
     class AnnotationPropertySpec:
         pass
 
-from ..util import annotation_property_specs, choose_output_spec
+from ..util import annotation_property_specs
+
+from ._util import _encode_uint64_series
+from ._id import _write_annotations_by_id
+from ._relationships import _write_annotations_by_relationships, _encode_relationships
+from ._spatial import _write_annotations_spatial
 
 logger = logging.getLogger(__name__)
 
@@ -295,19 +298,6 @@ def _encode_annotations(df, coord_space, annotation_type, property_specs, relati
     return df
 
 
-def _encode_uint64_series(s, dtype='<u8'):
-    """
-    Encode a pandas Series (or Index) of N values
-    into a numpy array of N buffers (bytes objects).
-    """
-    id_buf = s.to_numpy(dtype).tobytes()
-    id_bufs = [
-        id_buf[offset:(offset+8)]
-        for offset in range(0, len(id_buf), 8)
-    ]
-    return np.array(id_bufs, dtype=object)
-
-
 def _encode_geometries_and_properties(df, coord_space, annotation_type, property_specs):
     """
     For each annotation in the given dataframe, encode its geometry columns (e.g. x,y,z)
@@ -360,316 +350,3 @@ def _encode_geometries_and_properties(df, coord_space, annotation_type, property
     return ann_bufs
 
 
-def _encode_relationships(df, relationships):
-    """
-    For each annotation in the given dataframe, encode the related IDs
-    for all relationships into a buffer according to the neuroglancer spec.
-
-    Returns:
-        pd.Series of dtype=object, containing one buffer for each annotation.
-    """
-    if not relationships:
-        return None
-
-    encoded_relationships = {}
-    for rel_col in relationships:
-        encoded_relationships[rel_col] = _encode_related_ids(df[rel_col])
-
-    # Concatenate buffers on each row.
-    # Note:
-    #   Using sum() is O(N^2) in the number of relationships, but we generally
-    #   expect few relationships, so this is faster than df.apply(b''.join, axis=1)
-    rel_bufs = pd.DataFrame(encoded_relationships, index=df.index).sum(axis=1)
-    return rel_bufs
-
-
-def _encode_related_ids(related_ids):
-    """
-    Given a Series containing lists of IDs, encode each list of IDs
-    in the format neuroglancer expects for each relationship in an
-    annotation.
-
-    Each item in related_ids is a list, which gets encoded as
-    <count><id_1><id_2><id_3>..., where <count> is uint32 and
-    <id_1><id_2><id_3>... are each uint64.
-
-    Args:
-        related_ids:
-            A Series of length N and dtype=object, containing lists of IDs.
-            As a special convenience in the case where every row contains
-            exactly one ID, you may pass a series with dtype=np.uint64,
-            which will be interpreted as if each entry were a list of length 1.
-            (In this case, the implementation is slightly faster than in the general case.)
-
-    Returns:
-        A numpy array with N entries, where each entry is a buffer as shown above.
-    """
-    # Special case if the relationship contains only a single ID.
-    if np.issubdtype(related_ids.dtype, np.integer):
-        buf = (
-            pd.DataFrame({'count': np.uint32(1), 'id': related_ids})
-            .astype({'count': np.uint32, 'id': np.uint64}, copy=False)
-            .to_records(index=False)
-            .tobytes()
-        )
-
-        encoded_ids = [
-            buf[i*12:(i+1)*12]
-            for i in range(len(related_ids))
-        ]
-        return np.array(encoded_ids, dtype=object)
-
-    # Otherwise, the relationship contains lists.
-    else:
-        assert related_ids.dtype == object
-        counts = related_ids.map(len).to_numpy(np.uint32)
-        offsets = 8 * np.cumulative_sum(counts, include_initial=True)
-
-        ids_buf = np.concatenate(related_ids, dtype=np.uint64).tobytes()
-        counts_buf = counts.tobytes()
-
-        encoded_ids = [
-            counts_buf[i*4:(i+1)*4] + ids_buf[start:end]
-            for i, (start, end) in enumerate(zip(offsets[:-1], offsets[1:]))
-        ]
-        return np.array(encoded_ids, dtype=object)
-
-
-def _write_annotations_by_id(df, output_dir, write_sharded):
-    """
-    Write the annotations to the "Annotation ID Index", a subdirectory of output_dir.
-
-    Returns:
-        JSON metadata to be written under the 'by_id' key in the top-level 'info' file.
-        Currently, this is always {"key": "by_id"}
-    """
-    if 'rel_buf' in df.columns:
-        ann_bufs = df['ann_buf'] + df['rel_buf']
-    else:
-        ann_bufs = df['ann_buf']
-
-    logger.info("Writing annotations to 'by_id' index")
-    metadata = _write_buffers(ann_bufs, output_dir, "by_id", write_sharded)
-    return metadata
-
-
-def _write_annotations_by_relationships(df, relationships, output_dir, write_sharded):
-    """
-    Write the annotations to a "Related Object ID Index" for each relationship.
-    Each relationship is written to a separate subdirectory of output_dir.
-
-    Returns:
-        JSON metadata to be written under the 'relationships' key in the top-level 'info' file,
-        consisting of a list of JSON objects (one for each relationship).
-    """
-    by_rel_metadata = []
-    for relationship in relationships:
-        metadata = _write_annotations_by_relationship(
-            df,
-            relationship,
-            output_dir,
-            write_sharded
-        )
-        by_rel_metadata.append(metadata)
-
-    return by_rel_metadata
-
-
-def _write_annotations_by_relationship(df, relationship, output_dir, write_sharded):
-    """
-    Write the annotations to a "Related Object ID Index" for a single relationship.
-
-    Returns:
-        JSON metadata for the relationship, including the key and sharding spec if applicable.
-    """
-    logger.info(f"Grouping annotations by relationship {relationship}")
-    bufs_by_segment = (
-        df[['id_buf', 'ann_buf', relationship]]
-        .dropna(subset=relationship)
-        .explode(relationship)
-        .groupby(relationship, sort=False)
-        # Use b''.join() instead of 'sum' to avoid O(N^2) performance for large groups.
-        .agg({'id_buf': ['count', b''.join], 'ann_buf': b''.join})
-    )
-    logger.info(f"Combining annotation and ID buffers for relationship '{relationship}'")
-    bufs_by_segment.columns = ['count', 'id_buf', 'ann_buf']
-    bufs_by_segment['count_buf'] = _encode_uint64_series(bufs_by_segment['count'])
-    bufs_by_segment['combined_buf'] = bufs_by_segment[['count_buf', 'ann_buf', 'id_buf']].sum(axis=1)
-
-    logger.info(f"Writing annotations to 'by_rel_{relationship}' index")
-    metadata = _write_buffers(
-        bufs_by_segment['combined_buf'],
-        output_dir,
-        f"by_rel_{relationship}",
-        write_sharded
-    )
-    metadata['id'] = relationship
-    return metadata
-
-def _write_annotations_spatial(df, bounds, output_dir, write_sharded):
-    """
-    Write the annotations to the spatial index.
-    Currently, we only support a single spatial grid level,
-    resulting in a single annotation list.
-    """
-    # According to the spec[1]:
-    #   "For the spatial index, the annotations should be ordered randomly."
-    #
-    # This probably doesn't matter here since we're using a limit of 1,
-    # but let's go ahead and follow the spec.
-    #
-    # [1]: https://github.com/google/neuroglancer/blob/master/src/datasource/precomputed/annotations.md
-    logger.info("Shuffling annotations for spatial index")
-    df = df.sample(frac=1)
-
-    logger.info("Concatenating all annotation buffers for the spatial index")
-    count_buf = np.uint64(len(df)).tobytes()
-    all_annotations_buf = b''.join(df['ann_buf'])
-    all_ids_buf = b''.join(df['id_buf'])
-    combined_buf = b''.join([count_buf, all_annotations_buf, all_ids_buf])
-
-    # For now, just one big buffer.
-    logger.info(f"Writing annotations to spatial index")
-    key = '_'.join('0' for _ in bounds[0])
-    metadata_0 = _write_buffers(
-        pd.Series([combined_buf], index=[key]),
-        output_dir,
-        "spatial0",
-        write_sharded
-    )
-
-    metadata = [
-        {
-            **metadata_0,
-            "grid_shape": [1] * len(bounds[0]),
-            "chunk_size": np.maximum(bounds[1] - bounds[0], 1).tolist(),
-
-            # According to jbms:
-            #   Neuroglancer "subsamples" by showing only a prefix of the list of
-            #   annotations according to the spacing setting.  If you set "limit" to 1 in
-            #   the info file, you won't get subsampling by default.  If you want the
-            #   subsampling to do something reasonable, then you can randomly shuffle the
-            #   order in which you write the annotations.
-            # Source:
-            #   https://github.com/google/neuroglancer/issues/227#issuecomment-651944575
-            "limit": 1,
-        },
-    ]
-    return metadata
-
-
-def _write_buffers(buf_series, output_dir, subdir, write_sharded):
-    """
-    Write the buffers to the appropriate subdirectory of output_dir,
-    in sharded or unsharded format.
-
-    Args:
-        buf_series:
-            pd.Series of dtype=object, whose values are buffers (bytes objects).
-            The index of the series provides the keys under which each item is stored.
-
-        output_dir:
-            str
-            The directory into which the exported annotations will be written.
-
-        subdir:
-            str
-            The subdirectory into which the buffers will be written.
-
-        write_sharded:
-            bool
-            If True, write the buffers in sharded format.
-            If False, write the buffers in unsharded format, i.e. one file per item.
-
-    Returns:
-        JSON metadata for the written data, including the key (subdir)
-        and sharding spec if applicable.
-    """
-    if write_sharded:
-        return _write_buffers_sharded(buf_series, output_dir, subdir)
-    else:
-        return _write_buffers_unsharded(buf_series, output_dir, subdir)
-
-
-def _write_buffers_unsharded(buf_series, output_dir, subdir):
-    """
-    Write the buffers to the appropriate subdirectory of output_dir,
-    in unsharded format, i.e. one file per item.
-
-    The index of buf_series is used as the key for each item, after being
-    converted to a string (as decimal values in the case of integer keys).
-
-    Returns:
-        JSON metadata, always {"key": subdir}
-    """
-    # In the unsharded format, the keys are just strings (e.g. decimal IDs).
-    string_keys = buf_series.index.astype(str)
-    buf_series = buf_series.set_axis(string_keys)
-
-    # Since we're writing unsharded files, we could have just used
-    # standard Python open() and write() here for each key.
-    # Using tensorstore here is mostly just a matter of taste, but it will
-    # become useful if we ever support alternative storage backends such as gcs.
-    kvstore = ts.KvStore.open(f"file://{output_dir}/{subdir}/").result()
-
-    # Using a transaction here is not necessary, at least for plain files.
-    # I'm not sure if it helps or hurts, but it probably doesn't matter much
-    # for small datasets, which is presumably what we're dealing with if the
-    # user has chosen the unsharded format.
-    with ts.Transaction() as txn:
-        for segment_key, buf in tqdm(buf_series.items(), total=len(buf_series)):
-            kvstore.with_transaction(txn)[segment_key] = buf
-
-    metadata = {"key": subdir}
-    return metadata
-
-
-def _write_buffers_sharded(buf_series, output_dir, subdir):
-    """
-    Write the buffers to the appropriate subdirectory of output_dir,
-    in sharded format.
-
-    The index of buf_series is used as the key for each item,
-    after being encoded as a bigendian uint64.
-
-    Returns:
-        JSON metadata, including the output "key" (subdir) and sharding spec.
-    """
-    # When writing sharded data, we must use encoded bigendian uint64 as the key.
-    # https://github.com/google/neuroglancer/pull/522#issuecomment-1923137085
-    bigendian_keys = _encode_uint64_series(buf_series.index, '>u8')
-    buf_series = buf_series.set_axis(bigendian_keys)
-
-    shard_spec = choose_output_spec(
-        total_count=len(buf_series),
-        total_bytes=buf_series.map(len).sum(),  # fixme, might be slow
-        hashtype='murmurhash3_x86_128',
-        gzip_compress=True
-    )
-    spec = {
-        "driver": "neuroglancer_uint64_sharded",
-        "metadata": shard_spec.to_json(),
-        "base": f"file://{output_dir}/{subdir}",
-    }
-    kvstore = ts.KvStore.open(spec).result()
-
-    # Note:
-    #   At the time of this writing, tensorstore uses a
-    #   surprising amount of RAM to perform the writes.
-    with ts.Transaction() as txn:
-        for segment_key, buf in tqdm(buf_series.items(), total=len(buf_series)):
-            kvstore.with_transaction(txn)[segment_key] = buf
-
-    metadata = {
-        "key": subdir,
-        "sharding": shard_spec.to_json()
-    }
-    return metadata
-
-
-def _assign_spatial_grid(df, coord_space, annotation_type, lower_bound, upper_bound, max_grid_depth):
-    """
-    Assign each annotation to a spatial grid cell.
-    """
-    grid_shape = np.ceil((upper_bound - lower_bound) / 2**max_grid_depth)
-
