+++ neuclease/misc/neuroglancer/annotations/util.py
@@ -1,5 +1,3 @@
-from dataclasses import dataclass
-from typing import Literal
 from collections.abc import Mapping, Collection
 
 import numpy as np
@@ -139,83 +137,3 @@ def _proptype(s):
     if (s.map(len) == len("#rrggbbaa")).all():
         return 'rgba'
     raise RuntimeError("Not valid RGB or RGBA colors")
-
-
-@dataclass
-class ShardSpec:
-    type: str
-    hash: Literal["murmurhash3_x86_128", "identity_hash"]
-    preshift_bits: int
-    shard_bits: int
-    minishard_bits: int
-    data_encoding: Literal["raw", "gzip"]
-    minishard_index_encoding: Literal["raw", "gzip"]
-
-    def to_json(self):
-        return {
-            "@type": self.type,
-            "hash": self.hash,
-            "preshift_bits": self.preshift_bits,
-            "shard_bits": self.shard_bits,
-            "minishard_bits": self.minishard_bits,
-            "data_encoding": str(self.data_encoding),
-            "minishard_index_encoding": str(self.minishard_index_encoding),
-        }
-
-
-def choose_output_spec(
-    total_count,
-    total_bytes,
-    hashtype: Literal["murmurhash3_x86_128", "identity_hash"] = "murmurhash3_x86_128",
-    gzip_compress=True,
-):
-    """
-    Copied from Forrest Collman's PR:
-    https://github.com/google/neuroglancer/pull/522
-    """
-    import tensorstore as ts
-    MINISHARD_TARGET_COUNT = 1000
-    SHARD_TARGET_SIZE = 50000000
-
-    # if total_count == 1:
-    #     return None
-    # if ts is None:
-    #     return None
-
-    # test if hashtype is valid
-    if hashtype not in ["murmurhash3_x86_128", "identity_hash"]:
-        raise ValueError(
-            f"Invalid hashtype {hashtype}."
-            "Must be one of 'murmurhash3_x86_128' "
-            "or 'identity_hash'"
-        )
-
-    total_minishard_bits = 0
-    while (total_count >> total_minishard_bits) > MINISHARD_TARGET_COUNT:
-        total_minishard_bits += 1
-
-    shard_bits = 0
-    while (total_bytes >> shard_bits) > SHARD_TARGET_SIZE:
-        shard_bits += 1
-
-    preshift_bits = 0
-    while MINISHARD_TARGET_COUNT >> preshift_bits:
-        preshift_bits += 1
-
-    minishard_bits = total_minishard_bits - min(total_minishard_bits, shard_bits)
-    data_encoding: Literal["raw", "gzip"] = "raw"
-    minishard_index_encoding: Literal["raw", "gzip"] = "raw"
-
-    if gzip_compress:
-        data_encoding = "gzip"
-        minishard_index_encoding = "gzip"
-
-    return ShardSpec(
-        type="neuroglancer_uint64_sharded_v1",
-        hash=hashtype,
-        preshift_bits=preshift_bits,
-        shard_bits=shard_bits,
-        minishard_bits=minishard_bits,
-        data_encoding=data_encoding,
-        minishard_index_encoding=minishard_index_encoding,
-    )
